requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
speedtest-cli
apscheduler
